:ok_hand: Move some function following code review
