+++ tests.py
@@ -12,6 +12,72 @@ from utils import (predict_references, predict_structure, load_csv_file,
 from settings import settings
 
 
+def test_get_reference_components(predicted_number_refs, actual_number_refs):
+    """
+    Input:
+    - How many references found in each document
+    - How many references there actually were in a sample of documents
+    Output:
+    - How much these values match for the sample manually counted
+    - This is measured as the average difference as a proportion of the actual
+      number
+    - The comparison numbers
+    """
+
+    print("How well the number of references are predicted being computed...")
+    comparison = []
+    for index, row in actual_number_refs.iterrows():
+
+        # This will happen when there was a references section pulled only
+        # There are a few in the sample who had no references section pulled
+
+        this_document_predicted = predicted_number_refs.loc[
+            predicted_number_refs['uri'] == row['uri']
+        ].reset_index()
+
+        if (len(this_document_predicted) == 0):
+            this_document_predicted = None
+            difference_per = None
+        else:
+            this_document_predicted = this_document_predicted[
+                'Predicted number of references'
+            ][0]
+
+            if (row['Number of References in WHO pdf'] == 0):
+                difference_per = None
+            else:
+                difference_per = abs(
+                    row['Number of References in WHO pdf']
+                    - this_document_predicted
+                ) / row['Number of References in WHO pdf']
+
+        comparison.append(
+            [row['Number of References in WHO pdf'],
+             this_document_predicted, difference_per]
+        )
+
+    comparison = pd.DataFrame(
+        comparison,
+        columns=[
+            "Number of References in WHO pdf",
+            "Predicted number of references",
+            "Difference as proportion of actual number"
+        ]
+    )
+
+    mean_difference_per = np.mean(
+        comparison["Difference as proportion of actual number"]
+    )
+
+    print(
+        "Average difference between actual and predicted number of references",
+        "as proportion of actual number is ",
+        str(round(mean_difference_per, 4))
+    )
+
+    return mean_difference_per, comparison
+
+
 def split_string(publications, column_name, replace_strings_list=[(',', '.')]):
 
     publications_column = publications[publications[column_name].notnull()]
