+++ utils/__init__.py
@@ -1,19 +1,16 @@
 from .separate import (process_reference_section,
                        summarise_predicted_references,
-                       test_get_reference_components,
                        split_sections,
                        split_reference)
-from .predict import predict_references, predict_structure, test_structure
+from .predict import predict_references, predict_structure
 from .fuzzymatch import FuzzyMatcher
 from .loader import load_csv_file, load_json_file, load_pickle_file, get_file
 
 __all__ = [
     process_reference_section,
     summarise_predicted_references,
-    test_get_reference_components,
     predict_references,
     predict_structure,
-    test_structure,
     FuzzyMatcher,
     load_csv_file,
     load_json_file,
