+++ utils/predict.py
@@ -1,7 +1,4 @@
 import pandas as pd
-import numpy as np
-from sklearn.feature_extraction.text import TfidfVectorizer
-from sklearn.metrics.pairwise import cosine_similarity
 
 
 def predict_reference_comp(mnb, vectorizer, word_list):
@@ -243,125 +240,3 @@ def predict_structure(reference_components_predictions,
 
     print("Reference structure predicted")
     return all_structured_references
-
-
-def test_structure(predicted_reference_structures,
-                   actual_reference_structures):
-    """How much does each reference component prediction match the sample given
-    (actual_reference_structures)? Don't want to rely on document and reference
-    number for matching since these could change based on earlier parts of
-    the algorithm.
-    """
-    # Go through predicted_reference_structures and find the relevant row in
-    # actual_reference_structures
-    # If it exists then check how much each of the reference components match
-    print("Reference structures being tested ... ")
-    category_types = [
-        'Title',
-        'Authors',
-        'Journal',
-        'PubYear',
-        'Volume',
-        'Issue',
-        'Pagination',
-    ]
-    similarity_score = pd.DataFrame(columns=category_types)
-    # Will get updated if it exists
-    mean_similarity_score = None
-    if not predicted_reference_structures.empty:
-        # Go through all the predicted references and compare the ones which
-        # ones were in the sample of actual references
-        for index, reference in predicted_reference_structures.iterrows():
-            # Which actual references are from the document this predicted
-            # reference is from?
-            actual_document_references = actual_reference_structures.loc[
-                actual_reference_structures[
-                    'Document uri'
-                ] == reference['Document uri']
-            ]
-            # For the vast majority there will be none
-            if (len(actual_document_references) != 0):
-                ####
-                # 1. Which one actual reference matches highly with the
-                #    predicted reference title?
-                ####
-                vectorizer = TfidfVectorizer(
-                    lowercase=True,
-                    ngram_range=(1, 1)
-                )
-                # Which categories were there for this predicted reference?
-                category_exist = []
-                for cat in category_types:
-                    if cat in predicted_reference_structures.columns:
-                        if str(reference[cat]) != 'nan':
-                            category_exist.append(cat)
-                # Only try to find the similarity matches if there was a
-                # title predicted:
-                if 'Title' in category_exist:
-                    # Vectorize any of the categories that were predicted:
-                    if len(category_exist) >= 1:
-                        tfidf_matrix = vectorizer.fit_transform(
-                            [str(reference[cat]) for cat in category_exist])
-                    else:
-                        # It didn't predict any of the categories, but this
-                        # shouldn't happen I don't think
-                        print("error no categories predicted")
-                    # Vectorise the titles from the actual references
-                    actual_vectors = vectorizer.transform(
-                       actual_document_references['Title']
-                    )
-                    # Find the one actual reference that has the highest
-                    # title match
-                    # (Index will be 0 for the title one because it was first
-                    # in the list in category_types)
-                    title_similarity_score = [
-                        cosine_similarity(a, tfidf_matrix[0])
-                        for a in actual_vectors
-                    ]
-                    title_similarity_score_max = np.argmax(
-                        title_similarity_score
-                    )
-                    # Get the one actual reference which is most like the
-                    # predicted title
-                    actual_reference_max = actual_document_references.iloc[
-                        title_similarity_score_max
-                    ]
-                    ####
-                    # 2. Vectorise all the categories in this one actual
-                    #    reference
-                    ####
-                    actual_reference_max_vectors = vectorizer.transform([
-                        str(actual_reference_max[cat])
-                        for cat in category_exist
-                    ])
-                    # For each of the reference categories that existed for
-                    # this predicted reference how well does it match?
-                    similarity_score_max = [
-                        cosine_similarity(
-                            tfidf_matrix[cat],
-                            actual_reference_max_vectors[cat]
-                        )
-                        for cat in range(0, len(category_exist))
-                    ]
-                    similarity_score_max = [
-                        item2
-                        for sublist in similarity_score_max
-                        for sublist2 in sublist
-                        for item2 in sublist2
-                    ]
-                    # Save how well it did for each category:
-                    similarity_score_max = pd.DataFrame(
-                        [similarity_score_max],
-                        columns=category_exist
-                    )
-                    similarity_score = similarity_score.append(
-                        similarity_score_max
-                    )
-                    mean_similarity_score = similarity_score.mean()
-    print(
-        "Average similarity scores between predicted and actual references",
-        " for each component, using a sample of ",
-        str(len(actual_reference_structures)),
-        " references: \n" + str(mean_similarity_score)
-    )
-    return similarity_score, mean_similarity_score
